fix: address clippy warnings in solver and UI code
- topology.rs: remove redundant `y as i32` cast (unnecessary_cast)
- pbs_planner.rs: add #[allow(manual_flatten)] on intentional pattern
- rt_lacam/solver.rs: #[allow(dead_code)] on hash field, #[allow(map_entry)]
  on BFS loop, #[allow(unnecessary_unwrap)] on safe else-branch unwrap
- tpts.rs: #[allow(manual_swap)] on explicit index-based swap
- experiment/compact.rs: map_or(true, ...) → is_none_or(...)
- experiment/fullpage.rs: remove useless format!("RUN EXPERIMENT")

+++ src/core/topology.rs
@@ -485,7 +485,7 @@ pub fn parse_movingai_map(text: &str) -> Option<(GridMap, ZoneMap)> {
             }
             // '.' and 'G' are walkable. Everything else is obstacle.
             if ch != '.' && ch != 'G' {
-                obstacles.insert(IVec2::new(x as i32, y as i32));
+                obstacles.insert(IVec2::new(x as i32, y));
             }
         }
     }
